packaging: fix sudoers packaging
Previously sudoers file was distributed instead of sudoers.in.
But sudoers is a config file generated by configure and use
paths from the host which has built the tarball that may differ
from the ones where the package is installed.
Now sudoers.in is correctly packaged and sudoers correctly
generated during rpm creation or installation on host.

Change-Id: Ia5b7e548439d4444216942e0ee0ba567302770dc
Bug-Url: https://bugzilla.redhat.com/1013563
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>