agent: clean vdsm state if vm dies unexpectedly
If the hosted engine vm dies unexpectedly, vdsm will hold onto some
queryable state for the vm.  An attempt to start the vm will be met
with an error, "Virtual machine already exists".

To prevent this, check for VM state and attempt to clean it before a
startup attempt by using the vdsm 'destroy' verb.

Change-Id: I52c8f41c28fbc1942a8b392b275359df57c1b5ef
Bug-Url: https://bugzilla.redhat.com/1013638
Signed-off-by: Greg Padgett <gpadgett@redhat.com>