lib: fix acceptance of keyword arguments for vdsm api
A prior refactoring broke keyword argument usage in calls to the
run_vds_client_cmd() method.  This fix allows for both argument lists
and keyword arguments, and automatically converts the keyword arguments
to a dict that is accepted by the vdsm api.  Callers should be careful
to use the proper parameter form as required by the vdsm api.

Change-Id: I9931d6c87b0798152bf344ddd8b9328ce01565a8
Bug-Url: https://bugzilla.redhat.com/1015721
Signed-off-by: Greg Padgett <gpadgett@redhat.com>