agent: temporarily drop host score upon enexpected VM shutdown
When the agent detects that the engine VM is shut down unexpectedly,
it will lower the score to encourage the VM to start on a different
host.

This is an optimization to assist the user in putting a host
into maintenance mode more quickly.  If it is used in this manner, the
user should remember to enable host maintenance mode, else the score
will eventually revert to its usual number and the host may re-acquire
the engine VM.

Change-Id: If11d561a39cc9723ffdd55836db693fc2aad0575
Bug-Url: https://bugzilla.redhat.com/1015724
Signed-off-by: Greg Padgett <gpadgett@redhat.com>