agent, lib: agent compatibility with metadata updates
When the metadata version changes, both old and new agents should
recognize this and behave appropriately.  The desired behavior is
described in this comment from constants.py.in:

The range of valid metadata versions for this host is from
METADATA_PARSE_VERSION <= version <= METADATA_FEATURE_VERSION.  If an
incompatible change is made to the metadata, both the parse and feature
versions should be raised beyond the current feature version.  This will
ensure that a) the old agent will recognize that it has been superseded
and shut down the vm, and b) the newer agents will ignore the older
agents' reported scores, thus assuming control of the vm.

Change-Id: I2357ce26aeb20439dec161c7b8ee46d94a4b3e3d
Bug-Url: https://bugzilla.redhat.com/1036710
Signed-off-by: Greg Padgett <gpadgett@redhat.com>