Add support for email notifications
This adds a support for configurable email notifications.
Currently only state machine transitions generate a notification
event. The event is then processed by broker and email is
sent to recipients.

Events are filtered according to a config file broker.conf.
If the event type exists as a key in [notify] section and the
detail field contains the value assigned to that key (case insen-
sitive regex matching applies) then the email is sent.

Change-Id: Ibb466b8b130b75705f3cc9f817e377e79b3e81ba
Signed-off-by: Martin Sivak <msivak@redhat.com>