agent, broker: progressive score penalty for high cpu load
Penalize the host score progressively based on cpu load and duration
of that load.  The host load used doesn't count the load incurred by
the engine vm, in order to help keep host scores stable if the vm
changes hosts.

Additional related changes include not considering memory used by the
vm in the memory score, as well as logging changes to make it easier
to tell what the various factors were in generating the current score.

Change-Id: I9cb205f5eb3729a17692b8ab854c58ddfe862b2f
Bug-Url: https://bugzilla.redhat.com/1034205
Signed-off-by: Greg Padgett <gpadgett@redhat.com>