Encapsulate the storage metadata and add support for blockSD
This creates an encapsulation class that is responsible for
creation and usage of storage metadata directory.

It allows to specify arbitrary number of services and then
initializes the proper files and mappings for the broker
to use them.

The example usage is:

b = FilesystemBackend("uuid-of-the-storage-domain", "nfs")
b.create({
  "hosted-engine.lockspace": 1048576,
  "hosted-engine.metadata": 1048576
})
b.connect()

Supported storage types are: nfs, block and glusterfs

The type serves only as a hint, the storage domain has to exist
and be already connected for the above to work. The class does
autodetection of the SD type in all cases except glusterSD.

To force a clean metadata files pass force_new=True argument
to the create call.

Change-Id: Iaced4ac08936fc9314ff821343ce33d29a5897cf
Signed-off-by: Martin Sivak <msivak@redhat.com>