use monotonic time when checking the timeouts
When using the system time the timers are influenced by the time changes
done by user or ntp. So if the time is shifted backwards, the timers are
delayed which can lead to longer downtime of HA VM running the engine.
Using monotonic time makes the timers independent on the system time.

Change-Id: I9794beb4082f52d29835be5b6a182ab448c248f2
Bug-Url: https://bugzilla.redhat.com/1036683
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>