don't use direct io on lv based devices
In python the O_DIRECT flag doesn't work properly,
because it needs the buffer alligned to 512B to read into and
in python it's not possible to control the size so the attempt
to read from file using O_DIRECT fails with an os exception.
It's also doubtful if it ever worked as intended because the
code used direct_io only to read, but not write.

Change-Id: Ie5264e1757e43f22e860af3869495ebab15e0d85
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>