Treat migration state as Up and check engine health normally
We used to treat migration source state as Vm up Health bad.
That is not the correct report as the source Vm still works
properly during live-merge.

This caused a situation when a migration caused other hosts
to try to start the Vm and fail because of sanlock protection.

This moves migration source state to the group of states with
Vm up Health <query the status page> behaviour.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1093366
Change-Id: If3c813461909103b11266c902ccc51b2df2d0d36
Signed-off-by: Martin Sivak <msivak@redhat.com>