added EngineStarting state
when then VM with engine is starting it takes some time and goes
thru states "waitforlaunch" and "powering up". During the start up
the engine was marked as UnexpectedlyDown because it wasn't 'up'
and the host got score 0 which might lead to killing the VM while
it's starting and trying to start it on another host where the
same situation repeats.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1093638
Change-Id: Ia7e5bb28eeb839b684be74ea768265d9dab01175
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>