use vdsm to get the storage domain path
We used to list all the dirs mounted in /rhev/data-center/mnt to find the
path to HE storage domain. This causes problems when there is an ISO domain
which is not accessible - the agent gets stuck in blocking I/O until the mounted
domain is online again.

Change-Id: Ia32197ba888623d34e169d14010802c71625d5ea
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1140824
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>
(cherry picked from commit 06992438916c5422b54416227d0b528c19c00510)