Add metadata cleanup functionality
This patch adds a new command line arguments to the agent
service.

--host-id, --clean and --force-clean

The agent connects to vdsm, broker and storage as usual,
cleans its metadata block using 0x0 bytes and quits.

If the metadata block contains dirty status (crc does not
match or stopped is not True) the agent reports an
error and quits. The force-clean option overrides this.

The host-id option can be used to clean metadata of other
agent from another hosted engine enabled host (the operation
requires running broker and proper storage configuration).

This change also releases sanlock when the agent stops.

Change-Id: I4e0979bdfc45e7e07c9970994d98dfa1c9c94be0
Signed-off-by: Martin Sivak <msivak@redhat.com>