Lockspace reset tool support
A new method is added to the Client class.

reset_lockspace(force=False)

This method connects to the broker, retrieves stats
and verifies that all hosts reported stopped==True
and that the HA cluster is in global maintenance
mode.

It then uses sanlock to reinitialize the lockspace
used to protect agent id numbers.

Change-Id: I1c9d1c329ad56f288aac50077411778f99d194f0
Signed-off-by: Martin Sivak <msivak@redhat.com>