states, broker: Check for already starting VMs on other hosts
When two or more hosts try to take over when one host goes down, only
one of them can start the hosted engine VM. The others will fail to
acquire the storage lock.

The health monitor now detects when this is the case, so that the host,
which could not acquire the lock, can remove the VM which it could not
start and go back to the down state.

Change-Id: Ia5d321039801a9868f2b1232ce0acfdfe756b477
Bug-Url: https://bugzilla.redhat.com/1207634
Signed-off-by: Roman Mohr <rmohr@redhat.com>