Do not start services, let systemd take care of that
RHEL 6 did not have any proper service monitoring and so we did
it by ourselves. RHEL 7 is the only supported system now and we
have systemd available so now we have two programs that are trying
to monitor and restart vdsmd and sanlock (us and systemd).

This causes trouble when there is a valid reason for keeping them
down - like during upgrade - and so this patch changes the behaviour
to just monitor and wait.

No services will be started by the agent, the fact that something
is not running will be logged and the agent will try later. This
won't increment the FSM error counters either so the agent won't
ever quit while waiting (unless stopped properly or killed by
other error).

Change-Id: I1178aff984d68739e3d42b582589d376b56f45c1
Signed-off-by: Martin Sivak <msivak@redhat.com>