Use SYNC DIRECT IO everywhere to make sure all hosts see the same data
We did remove direct io for LVM based backends as Python did not
supported that properly (http://bugs.python.org/issue5396).

The issue with direct io is that it requires page aligned buffers
for writes and reads. Python can't ensure that.

This patch uses a very nice workaround from
http://www.alexonlinux.com/direct-io-in-python that proposed to
use mmap to get the buffer. MMAPped memory is page aligned and
Python can use it for reads and writes.

Change-Id: I31912300b27a2ae215e2462114012676b43cb017
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1277079
Signed-off-by: Martin Sivak <msivak@redhat.com>