Stop monitoring domain during maintenance / shutdown
The vdsm's domain monitor has to be stopped when owned by
the HA agent. The stop is important for upgrade reasons as
it releases sanlock leases that would otherwise prevent
sanlock from upgrading.
Increasing killproc timeout from default (3 seconds) to 90
seconds (systemd defaults) to have all the time to properly
remove the lease and stop monitoring before being killed
with -9.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1298461
Change-Id: I7761fe7e40cb4df6a942b70d2f3307c098828293
Signed-off-by: Martin Sivak <msivak@redhat.com>