upgrade: fix upgrade from partially upgraded condition
The upgrade procedure has to move the hosted-engine
configuration to the shared storage and it has to
disconnect the hosted-engine storage domain from its
bootstrap storage pool.
The two tasks can be performed by different attempts.

_move_to_shared_conf was connecting the bootstrap
storage pool leaving it connected for further steps.
_remove_storage_pool was assuming that it was
already connected.
But if the configuration volume is OK it skips
_move_to_shared_conf and, if the system has been
rebooted in the middle, _remove_storage_pool fails
cause the storage pool is not connected.

Connecting the storage pool if needed also from
_remove_storage_pool

Change-Id: Ibc5782d2f245129b2bb2f7e70f2558d820eb4267
Bug-Url: https://bugzilla.redhat.com/1292652
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>