upgrade: Add (un)persist verbs to sudo
Previously sudo calls to persist and unpersist failed, because the sudo
configuration for the vdsm user did not allow it.
With this patch the vdsm user is allowed to run persist and unpersist.

Fixing also suders file for changing owner and goup of the same
config file with hust a command after
https://gerrit.ovirt.org/#/c/48700

Change-Id: Ic0fc4abaaf78b4b010031eca2a0d138743932989
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1280310
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>
(cherry picked from commit 08c050a384ac03499172dd999e181d789a106bff)