upgrade: fixing for iSCSI additional hosts
On 3.5 -> 3.6 we need to call getVolumesList to
scan all the available volumes and check if some other
host already created the configuration volume.

After a reboot ovirt-ha-agent calls prepareImage
only on metadata, lockspace and if available on
the configuration image but not on the engine image.

So, only after a reboot and only on iSCSI, getVolumesList
will report an error regarding the engine image.

filtering out know images from the scan,
using metadata image to check the images path,
avoid breaking on getVolumesList errors

Change-Id: I724835de6498fc7098d03499b246fa69b5bf09fe
Bug-Url: https://bugzilla.redhat.com/1297457
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>