storage: validating NFS/gluster paths
Validating NFS/gluster paths with os.path.normpath
as VDSM does.
Try to cleanly disconnect a previously connected
storage server if the storage server is already mounted
on a wrong path. Explicitly cleanup
/var/run/vdsm/storage/{sd} since VDSM is not and on
the other side it also doesn't refresh the link
if they are already there resulting in broken links.
Asking the user to reboot if disconnectStorageServer fails
(it will clean the whole /var/run becouse is on tmpfs).

Change-Id: I98273e99f1b855a4d96be706d864842e2ed61e1f
Bug-Url: https://bugzilla.redhat.com/1300749
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>