storage: validating NFS/gluster paths
Validating NFS/gluster paths with os.path.normpath
as VDSM does.
Try to cleanly disconnect a previously connected
storage server if the storage server is already mounted
on a wrong path.
Asking the user to reboot if disconnectStorageServer fails
(it will clean the whole /var/run becouse is on tmpfs).

Change-Id: Ie7c8daa814dfdc3aaf3bae382e90b8d166839f40
Bug-Url: https://bugzilla.redhat.com/1300749
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>