storage: always pass a blank SP UUID in image.py
If we are upgrading from 3.5 in the config file spuuid
still point to the bootstrap storage pool so we must ignore it.
On normal flow we don't realize it since all the images are
already prepared by the previous 3.5 run but, it the user
reboots the host just after yum update, the prepareImage
will fail after the reboot and since the images are not prepared
nothing else happens.
Always forcing spuuid = blank there. This would not affect
host that are already at 3.6 since in that case we will have
for sure spUUID=00000000-0000-0000-0000-000000000000
under /etc/ovirt-hosted-engine/hosted-engine.conf

Change-Id: I7bed963c8bbc4aca82185cdd7072dda4f3f2286c
Bug-Url: https://bugzilla.redhat.com/1297457
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
(cherry picked from commit d93fceca78c78bb625ba98069d6eee13d8e25e92)