upgrade: trying reconstructMaster as a recovery action
In the upgrade procedure from 3.5 we call reconstructMaster
to set the temporary fake storage domain on PosixFS
as the master storage domain in order to be able to call
detachStorageDomain for the hosted-engine storage domain.

The upgrade procedure is not transactional and the
temporary fake storage domain on PosixFS will not survive
a reboot.

So, if the user reboots the host just after the
reconstructMaster but before the detachStorageDomain, we
end with a storage pool that cannot be connected
since its master storage domain doesn't exist anymore.

reconstructMaster can be used to try to revert to
3.5 status (bootstrap storage pool with just the
hosted-engine storage domain as its master storage domain)
before trying the upgrade procedure again.

Change-Id: I9bfc3461694348de78750a5d0eea0dc65a36d15e
Bug-Url: https://bugzilla.redhat.com/1292652
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>