storage: improved workaround for getImagesList on NFS
We know from the past that getImagesList wasn't properly working
on 3.6 on NFS StorageDomain when the host was not connected to a
storagePool (as it happened in hosted-engine scenario).

There was a workaround for that triggered by an explicit error
in getImagesList return value ({'status': {'message': 'list index out
of range', 'code': 100}}).

Now it seems that getImagesList has started to report success (code=0)
for this case, although the image list is still empty.

This patch improves the workaround to catch that.

Change-Id: I26ed9e823f60415b500b87f11e91128228b45dec
Bug-Url: https://bugzilla.redhat.com/1319721
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>