storage: correctly evaluate glusterfs SD canonical path
In the past we let the user specify a path for file based
(NFS or gluster) storage domains which includes a trailing '/'.
This caused other issues (double mount point when the engine
import it) and so we had a patch to evaluate the canonical path
and eventually fix the mount point in the configuration file
before trying again to connect it. Due to a bug in the evaluation
of the canonical path for gluster volumes, it was looping forever
on gluster deployed systems.
Improoving the the logging to make it more evident.

Change-Id: Ic7a877bb4386214044cf85de18f99b1b443222eb
Bug-Url: https://bugzilla.redhat.com/1327516
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>