Add video device if not present in the ovf store
Because the video device was not always reported correct in the past
there are ways to import the HE VM into ovirt-engine without the video
device. This means that it will not appear in the ovf store and
therefore we would not ask for it on the next VM restart.

Here a fallback video device configuration is provided in that case.
This is only 3.6 related. On 4.0 all involved components are fixed.

Change-Id: I8b412e2d93ba6269743257e93342ceaa47eb2299
Related-To: https://bugzilla.redhat.com/1326810
Related-To: https://bugzilla.redhat.com/1328921
Bug-Url: https://bugzilla.redhat.com/1339305
Signed-off-by: Roman Mohr <rmohr@redhat.com>