Fix refreshing vm conf
When refreshing the vm configuration on the host
we first try the ovf store, if no content was found
we try the shared storage vm.conf.
When we have the content we can refresh the locally
stored file.
The recent refactor broke this functionality, this patch
fixes it.

Change-Id: I281b42a6cb2c5cfdd5821cd50e3f2608d3ae8896
Bug-Url: https://bugzilla.redhat.com/1301681
Signed-off-by: Jenny Tokar <jtokar@redhat.com>
(cherry picked from commit d771cf1)