Add EngineMaybeAway state to wait before EngineUnexpectedlyDown state
When the hosted engine VM is migrated quickly, the source agent will miss
the migation process and switch to EngineUnexpetedlyDown state.

This patch adds an intermediate state, EngineMaybeAway, which waits for the
VM to appear on another host. If it does not appear in a certain time,
it moves to EngineUnexpetedlyDown.

Change-Id: I1a71997bf744666a9b2cb6dd926d034a945d76a0
Bug-Url: https://bugzilla.redhat.com/1419326
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>