Add graphics device to the vm.conf
When an ovf file was converted to the vm.conf file
the graphics device was not converted.
If the vm was started with a vm.conf file that didn't
contain the graphics device there was no option to
connect to the graphical console from the engine.
This patch fixes this behavior.

Change-Id: I3f3c3f59a02a9e78baa894c2e4a668face0f64f4
Bug-Url: https://bugzilla.redhat.com/1441570
Signed-off-by: Jenny Tokar <jtokar@redhat.com>