Add transition to EngineMaybeAway from EngineStarting
When the agent tries to start the VM and it is already
running on another host, the VDSM reports an error:
"Failed to acquire lock".

Broker periodically checks the state of the VM,
but it sees the error message only once and then
the VM is destroyed.
The agent may miss the message if it sleeps too long.

This patch adds transition to EngineMaybeAway,
to check if the VM is running elsewhere.

Change-Id: I4d92e021ebf6398d7846a233531f1be5e94dbe5f
Bug-Url: https://bugzilla.redhat.com/1448699
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>