packaging: fixing bad permissions on answer file from the past
answers.conf is in the configuration volume on the shared
storage since 3.6 while it was just a local file before.
The 3.5 -> 3.6 upgrade procedure is able to move it to
the shared storage but it runs in ovirt-ha-agent deamon
as vdsm user so answers.conf should be readable by vdsm user.
This patch has been previously applied to
ovirt-hosted-engine-setup but it's more correct to have it here
since the upgrade is performed by ovirt-ha-agent.

Change-Id: Iddc518e4f326bbc4831843e0b980348eb3ebafa3
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>