logging: Loop execution time is useful thing and should be logged as info.
Adjustable log level was only used to skip report loop execution time and delays
and can be safely replaced with static values. Loop execution time is
a interesting metric, so it should be logged at info level all the time and
there is not need to keep log level adjustable for only a single log message.

Bug-Url: https://bugzilla.redhat.com/1378310
Change-Id: I4e6441586c7c1e8d1f57e7bdb8d12867844d0912
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>