broker: Broker should read storage info from the config.
As we do not support multiple storages now, it makes no sense
to accept storage info from the agent. Therefore, broker must be
able to get storage info (and construct relevant objects) on it's own,
using a config file. This also makes set-storage RPC call useless,
so it is removed.

This also removes support for different storage backends in broker,
so FilesystemBackend is not supported anymore.

https://github.com/oVirt/ovirt-site/pull/1157

Change-Id: I7dec5f52689d5051af4e2ec0e56401fd65fa1d6b
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>