storage: Storage must be fully initialized by the broker.
With stateless RPC broker may not rely on agent to configure storage
and must do it on its own. This patch copiess part of storage initializaion
code from agent to the broker, to enable that functionality. In a future
we plan to move all storage handling code from agent to the broker.

Change-Id: I77ceb08ab1a33384e7470a1c4a40c38acf1039a1
Bug-Url: https://bugzilla.redhat.com/1488333
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>