iscsi: multipath: connect the whole portal group
Connect all the portals of a portal group to enable
iSCSI multipath.
To be as backward compatibile as possible and avoid
requiring changes in host-deploy, it saves the
multiple IP and port values as string composed
by multiple comma separated values with the
same key used by previous version.
It assumes that the login is the same for all
the portals.

Change-Id: I012743c07d36ec79e6dd17fa6e8248f5648dc57e
Bug-Url: https://bugzilla.redhat.com/1149579
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>