iscsi: multipath: support multiple iSCSI interfaces
Parse iSCSI interfaces configuration files from
/var/lib/iscsi/ifaces and try to add a path for
each interface there to all the configured portals.
This approach could lead to unneeded failed paths
since ovirt-ha-agent will try to connect the
whole cartesian product of portals and interfaces while
the engine can eventually filter it.
A new value in he configuration file,
iscsi_paths_blacklist, is introduced to let the user
specifiy a blacklist of paths to be skipped by
ovirt-ha-broker as well.
iscsi_paths_blacklist could be set just locally on the
host with --type=he_local or with --type=he_shared to
be consumed also by the host deployed after that.
If needed, a new warning line will appair in the log
suggesting to the user to set iscsi_paths_blacklist to
speed up future boostrap.
The syntax for iscsi_paths_blacklist is a comma
separated lists of 'interface<>portal_ip' couples.
example:
 iscsi_paths_blacklist=eth2<>192.168.2.125,eth2<>192.168.1.125

Change-Id: I659b9e754f25206565e7989d5fa74ed165faad56
Bug-Url: https://bugzilla.redhat.com/1193961
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>