Broker: Release lock in engine_health monitor when checking engine liveliness.
When engine_health monitor is checking for engine liveliness and the VM is shut down,
the call will have to time out, which can take a longer time.
Meanwhile an event that the VM was shut down is received, but not processed until
the liveliness check times out.

This patch allows the event to be processed while waiting for the liveliness check.

Change-Id: I88be90fe690fda34f133308938c2981aaa17fa56
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>