Do not crash the agent on notification failure
There was a chance of agent crashing when broker became unresponsive. This
will happen less when all the broker refactoring patches are merged,
but it does not hurt to be a bit more defensive here.

It does not directly solve the references bug either as the reason
for unresponsive broker needs to be solves. This will just remove one
of the symptoms.

Bug-Url: https://bugzilla.redhat.com/1413845
Change-Id: Icd41ab75131489106020b7505cc3d5b9ed20bcf0
Signed-off-by: Martin Sivak <msivak@redhat.com>