agent: Detect canceled migration
When a VM migration fails, the vdsm RPC call getMigrationStatus returns
an error status and progress 0%. The HE ignores the status code and
continues as if the migration was still in progress.

This patch checks that the VM is in one of 2 states while migrating:
- MIGRATION_SOURCE - migration is in progress
- DOWN - migration succeeded

Any other state is a migration failure.

Change-Id: Ib09ecfe0bdb49fac0f288aab5553bbf82b6cadfc
Bug-Url: https://bugzilla.redhat.com/1519289
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>