Prepare a volume symlink again if it disappeared
This is a workaround for a situation where engine reads
the hosted engine configuration for the purpose of deploying
an additional HE host. The engine then instructs vdsm
to teardown the volume.

VDSM does not have any reference count tracking and so it
does not know the host is still using it as well.

Bug-Url: https://bugzilla.redhat.com/1527394
Change-Id: I356283407704494d01f1d8955f90372e11a8355f
Signed-off-by: Martin Sivak <msivak@redhat.com>