storage: avoid preparing all the images in advance
Avoid preparing all the images in advance
at VdsmBackend.connect time.
Now all the code is robust enought to detect
that a disk is not ready and
call prepareImage on the fly just if needed
(the engine could call teardownImage on OVF_STORE
disks or configuration disk for different reasons).
In 4.2 node-zero flow there is no need to
add a regular storage domain to become the
master storage domain so the user can
end creating many VMs on the hosted-engine
storage domain and so activating all
the LVs in advance could be more risky.

Change-Id: I5b5301619edec6c320d86118ed93569d6e128da8
Bug-Url: https://bugzilla.redhat.com/1443819
Requires: https://gerrit.ovirt.org/87326/
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>