agent: Add local maintenance manual mode
The HA agent needs to be aware of manual mode of local maintenance
in addition to the existing local maintenance mode.

When setting the maintenance mode:
local maintenance = "local maintenance" OR "local maintenance manual"

Local maintenance manual mode is set to True
when invoking the hosted engine tool for enabling maintenance:
hosted-engine --set-maintenance --mode=local
It will change to False when --mode is set to none or global

Its default value is False.

Change-Id: Ia14298f8d7b4a62747850558de28fa17fbe25c91
Bug-Url: https://bugzilla.redhat.com/1469143
Signed-off-by: Yanir Quinn <yquinn@redhat.com>