broker: restart the service on storage errors
The engine could disconnect the hosted-engine SD
due to maintenance mode or similar.
Restart the broker if it fails to read or write
status from metadata area to force a fresh
storage connection.
In the past this behaviour was in charge of
the agent, and the code is still there,
but https://gerrit.ovirt.org/82591/
moved the storage initialization to the
broker and so restarting the agent is
useless in that cases.

Change-Id: I2237d725e0e43d98e85239a75209efb63171ddfa
Bug-Url: https://bugzilla.redhat.com/1583712
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
(cherry picked from commit 2e859f1ab08a19021f6e6871e6ff338292687856)