text: rephrase brokerlink exceptions
Now if ovirt-ha-broker is down, for instance
due to a storage issue, the agent will detect it
due to the missing unix domain socket file.
This will result in something like
  RequestError: Failed to start monitor ping, options {'addr': '10.70.43.254'}: [Errno 2] No such file or directory

on ovirt-ha-agent logs that can be
misunderstood as network issue
sending a ping.
Rephrasing it as:
  RequestError: brokerlink - failed to start monitor via ovirt-ha-broker: [Errno 2] No such file or directory, [monitor: ping, options: {'addr': '172.30.1.254'}]
to make more evident that the issue is just in
the communication channel to ovirt-ha-broker.

Change-Id: Ib15d942acd20f9cc7a335b13153feba1af786ffc
Bug-Url: https://bugzilla.redhat.com/1639997
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>