Exit unconditionally on errors after some timeout
It has been experienced that the broker stopped responding to some
requests.  The following scenario is considered: While the listener
may no longer process requests, it may still listen on its socket.
Connections to the broker may be initiated but they never get any
response and they hang.  The broker is then unavailable, clients don't
get requested information and they may get stuck, e.g. hanging Vdsm
executor workers connecting to the broker may log tracebacks and/or
block the executor.

While this is expected to happen only in erroneous situations or
broken setups, the broker should still be resistant against that.
The easiest way is to make it stop in a harsh way in case it doesn't
stop gracefully in some reasonable time.

Change-Id: I068f2131806800a179c908b0fa0fc54c996c1207
Bug-Url: https://bugzilla.redhat.com/1569593
Bug-Url: https://bugzilla.redhat.com/1596952
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>