Fix unicode handling on python-3 on EngineXml
The OVF generated by the engine contains
an element called EngineXml which is the
XML VM definition for libvirt.

The external (wrapper) OVF file is encoded
in UTF-8 as specified in its XML header;
the code is reading it with 'rb' mode passing
the bytes to lxml parser that
correctly handle their encoding due to the header.

Inside that file we have EngineXml element which
looks like:
<EngineXml>&lt;?xml version="1.0" encoding="UTF-8"?&gt;
lxml parser correctly decode it returnign an
unicode child object that we have to encode again
to UTF-8 (as specified in the header of the nested XML)
before being able to parse it again with lxml that
will decode it according to its header
(so UTF-8 again).

Change-Id: I4199ccf500b53785f31e16f5ef1f6644ce7770d7
Bug-Url: https://bugzilla.redhat.com/1757414
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>