agent: On EngineUp and vm down, go to EngineMaybeAway
This is often a result of an engine vm migration, where the destination
host still didn't manage to update the shared storage saying the vm is
up.

So instead of trying to start it immediately (and hopefully/usually
failing due to sanlock), move to EngineMaybeAway, to give it some more
time.

Bug-Url: https://bugzilla.redhat.com/1815589
Change-Id: I1c7def9264c287abe5b1bf746c1e30f6577754de
Signed-off-by: Yedidyah Bar David <didi@redhat.com>