Increase MAX_HOST_ID_SCAN to 2000
Before this patch, it was 64, which allowed Host IDs to be up to 64.
In practice, it was enough most of the time, but removing a host does
not "free" its ID, so a long-running setup with many host additions and
removals can reach IDs higher than that, even without having actual 64
hosts.

The shared storage has space for much more than that.

sanlock allows up to 2000 hosts to use its lockspace, so we are limited
by that anyway, re the number of hosts.

This only slightly increases memory and time usage when reading/parsing
the storage - by my measurements, it adds something like up to 30
miliseconds per read (which happens usually once every 10 seconds).

Bug-Url: https://bugzilla.redhat.com/1916032
Change-Id: I33b2fcd5d560a45c73773570e2b0508dfe61e541
Signed-off-by: Yedidyah Bar David <didi@redhat.com>