python3: broker/listener.py: Silence _encode
I spent quite some time trying to understand when we pass bytes and not
a string, and failed. Give up for now and stop spamming the log - no-one
seems to look at it or care, and I never noticed a relation to an actual
problem/failure - although this does not mean there is no such relation,
but I hope the existing logging will be enough.

Change-Id: I8a58f002996f00abc8dcb7dad1a429c862feaa1b
Signed-off-by: Yedidyah Bar David <didi@redhat.com>